LayoutTests/fast/loader/loadInProgress-expected.txt

ALERT: Test passed: [object FORM]
